+++ packages/iov-tendermint-rpc/src/responses.ts
@@ -284,11 +284,11 @@ export interface Validator {
 }
 
 export interface ConsensusParams {
-  readonly blockSize: BlockSizeParams;
+  readonly block: BlockParams;
   readonly evidence: EvidenceParams;
 }
 
-export interface BlockSizeParams {
+export interface BlockParams {
   readonly maxBytes: number;
   readonly maxGas: number;
 }